+++ src/hklpy2/__init__.py
@@ -65,6 +65,7 @@ from .diffract import diffractometer_class_factory  # noqa: E402, F401, F403
 from .incident import A_KEV  # noqa: E402, F401
 from .exceptions import SolverError  # noqa: E402, F401
 from .run_utils import ConfigurationRunWrapper  # noqa: E402, F401
+from .run_utils import register_aux_reconstructor  # noqa: E402, F401
 from .run_utils import simulator_from_config  # noqa: E402, F401
 from .run_utils import get_run_orientation  # noqa: E402, F401
 from .run_utils import list_orientation_runs  # noqa: E402, F401